Fix documentation
This commit fixes multiple issues with documentation.

1. Duplicated requirejs

The nbsphinx extension introduced in #2393 pulled a requirejs
which caused the initialization script to halt.

As a result, the right side bar was left uninitialized.

2. Undefined variable error

It turned out that PyTorch's theme expected the downstream projects
to define `collapsedSections` variable.

Currently console log shows `collapsedSections is not defined`.
As a result of this fix, we start to see the + symbol on left side.

3. Fix the behavior of default expand

Tweaks the right-side bar initialization behavior
so that expand-all only happens once, not at every resize.

4. Overwrite the link to GitHub

The `GitHub` tab in main-menu always linked PyTorch core.
This commit adds overwrite to torchaudio page