Make FFmpeg log level configurable (#2439)
Summary:
and set the default level to 8 for less verbose logging.

Undesired logs are one of the loudest UX complains we get.
Yet, loading media files involves uncertainty which is
difficult to debug without debug log.

This commit introduces utility functions to configure logging level
so that we can ask users to enable it when they encounter an issue,
while defaulting to non-verbose option.

Pull Request resolved: https://github.com/pytorch/audio/pull/2439

Reviewed By: xiaohui-zhang

Differential Revision: D36903763

Pulled By: mthrok

fbshipit-source-id: ed495641a868ed9a3f34c2a5f37fa3e5b135d0ec