Tweak badge link URL generation
Currently, the way feature badges are generated assumes that
both documentations and the supported features page are on the
same level from the root.

This does not work when we introduce `:autosummary:` which generates
individual documentation pages one level below.

This commit changes it so that links to the supported features page
are properly relative from the documentation level.

There is no appearance change from this commit.