Tweak badge link URL generation (#2677)
Summary:
Currently, the way feature badges are generated assumes that both documentations and the supported features page are on the same level from the root.

This does not work when we introduce `:autosummary:` which generates individual documentation pages one level below.

This commit changes it so that links to the supported features page are properly relative from the documentation level.

There is no appearance change from this commit.

Pull Request resolved: https://github.com/pytorch/audio/pull/2677

Reviewed By: carolineechen

Differential Revision: D39507451

Pulled By: mthrok

fbshipit-source-id: f18da4201f0eb747586be21c8bd9a958217aebc2