Fix leaking matplotlib figure
In StreamWriter basic usage tutorial, matplotlib is used to generate
raster images of waveforms, and the figure used is left unshown in
the resulting tutorial with the use of ``sphinx_gallery_defer_figures``
command.

It turned out that this figure is shown in the next code block executed
by Sphinx Gallery, and the figure is placed in totally unrelated place.

This commit fixes it by closing the figure.