Fix the handling of discard_before_pts
Currently `discard_before_pts=-1` is used to indicate no AVFrame should be skipped.
It was reported that some corrupted video can have constant negative pts value.

It is technically UB for such corrupted data, but still all the AVFrame should be
decoded as long as `seek` is not used.

This commit changes the decoder so that it processes AVFrame if `discard_before_pts==-1`
disregard of AVFrame::pts value.