Fallback to best_effort_timestamp in case of invalid PTS
If the input video has invalid PTS, the current precise seek fails
except when seeking into t=0.

This commit updates the discard mechanism to fallback to
`best_effort_timestamp` in such cases.

`best_effort_timestamp` is just the number of frames went through
decoder starting from the beginning of the file.

This means if the input file is very long, but seeking towards the
end of the file, the StreamReader still decodes all the frames.

For videos with valid PTS, `best_effort_timestamp` should be same as
`pts`.