Split extract_archive into dedicated functions.
`extra_archive` in `datasets.utils` does not distinguish
the input type, and blindly treats it as tar, then zip in case of failure.

This is an anti-pattern. All the dataset implementations know
which archive type the downloaded files are.

This commit splits extract_archive function into dedicated functions,
and make each dataset use the correct one.