Add multithread options to docstring and default to single thread
One can pass "threads" and "thread_type" to `decoder_option` of
StreamReaader to change the multithreading configuration.

These affects the timing that decoder starts emitting the decoded
frames. i.e. how many packets at minimum have to be processed before
the first frame is decoded.

Overall, multithreading in decoder does not improve the performance.
This is because the design of StreamReader, decode few frames then
fetch them, does not suited to saturate the decoder with incoming
packets.

For this reason, we default to single thread execution in StreamReader.

Follow-up: Apply similar change to encoder option in StreamWriter.