Merge pop_chunks methods (#3002)
Summary:
Pull Request resolved: https://github.com/pytorch/audio/pull/3002

This commit merges `pop_chunks` and `pop_chunks_with_metadata`.

In #2975 (D42526945 (https://github.com/pytorch/audio/commit/0dd59e0dda22eabf54fc95ad8050094df239bd39)), we updated StreamReader so that it returns PTS.
In that PR, we introduced `pop_chunks_with_metadata` method, so that
the original `pop_chunks` method returns the same type and we could
focus on the PTS logic in the code review.

The commit is landed, now we merge the two methods, so that the original
`pop_chunks` returns Tensor frames and metadata (PTS).

Reviewed By: xiaohui-zhang

Differential Revision: D42662321

fbshipit-source-id: 37ae088bc63fc516ea068698088925e8b31bc0a1