Add get_build_config ffmpeg utility function
We often need to look at which FFmpeg was found and linked when
debugging an issue.

Version number is often not enough but there is no easy way to find
where the library was found either.

This commit adds utility function that prints the build time
configuration.

It helps to distinguish if the linked FFmpeg is the one from binary
distribution built in CI or locally built.