Remove unnecessary AVFrame allocation
Summary:
When input format and encode format is different in StreamWriter, filter for format conversion is inserted.

A temporary AVFilter (`dst_frame`) is used for this case,
but FilterGraph handles the memory allocation,
so there is no need to perform allocation by ourselves.

This `dst_frame` is otherwise not used, so we do not have to allocate memory at all.
This commit removes the unnecessary memory allocation at all.

Differential Revision: D42865042

fbshipit-source-id: 181c5a6de4f7cacd914b28cb0bc09eebcbbe6fb0