Enhance UX on TorchAudio pages to improve awareness of doc versioning
- Boldface the version-selection UX and increase size by three percent.
- Add text to breadcrumbs to indicate version and stability.
- New `breadcrumbs.html` in `_templates` overrides Sphinx version.

I create a new variable in conf.py, `version_stable`, which has the
version number for the most-recent stable release. I define this
variable in the `html_context` dictionary so that it is visible to the
templates.

I use this approach because I was not able to find any other
way of discerning the current stable release during the build. Note that
the versions.html file--which identifies the current stable
release--appears to be available only in the gh-pages branch and so it
is not available at build time.

However, this means that someone will need to update conf.py whenever
the current stable release changes.