Enable strict=experimental automatically when encoding opus/vorbis
OPUS encoder and VORBIS encoders require "stric=experimental" flags.
This commit enables it automatically.

The rational behind of it is typically we care if we can encode these
formats at all and not how they are encoded. (This might be concern
when these encoder becomes more mature on FFmpeg side and providing
flags would result in weird behavior)

Also when writing high-level functions that uses StreamWriter, if we
do not set these flags, then these high-level functions have to add
new options that should be passed down to StreamWriter, which turned
out to be very painful in #1363