Warn if decoding YUV images with different plane size (#3201)
Summary:
StreamReader behaves differently when dealing with YUV formats.
It implicitly converts the image format to YUV444P because
otherwise image planes do not have the same shape and it is not
possible to express it as a regular PyTorch Tensor with dedicated
dimension for each color channel.

This is commit adds warnings to such conversions.

Pull Request resolved: https://github.com/pytorch/audio/pull/3201

Reviewed By: nateanl

Differential Revision: D44311017

Pulled By: mthrok

fbshipit-source-id: c492d51cfba77a346a7afc30f4edf3243a28bf8b