Split decoder and dst stream configuration logic in StreamReader (#3213)
Summary:
Pull Request resolved: https://github.com/pytorch/audio/pull/3213

Currently, `StreamReader` expects users to add an output stream using `add_audio_stream`. The function, however, is responsible for configuring both the decoder for the source stream and the transforms to be applied on the decoded frames to produce the output stream. In the scenario where the user adds multiple output streams to a given source stream, this can be confusing, since the user is able to supply decoder options in each invocation, but only those that were provided in the first invocation will be used to configure the decoder.

To better reflect the split between the decoder (source stream) and frame processors (output streams) and to improve usability, this PR introduces separate source stream and destination stream configuration functions. The expectation here is for the user to first configure the source stream and then add destination streams to that source stream.

Differential Revision: D44448429

fbshipit-source-id: 6b14f59c25a21b8d2b5e9961dd482a87a84c25e7